requirements.txt
fastapi
uvicorn 
python-dotenv
azure-search-documents
openai
pydantic
pdfplumber
python-docx
tiktoken 
streamlit
